requirements.txt
# Noise Generator Dependencies
opencv-python>=4.5.0
numpy>=1.20.0
matplotlib>=3.3.0
scikit-image>=0.18.0
tqdm>=4.60.0
pathlib
rasterio>=1.3.0
Pillow>=8.0.0
